Refactor: lsm/store core API and lsm iterators
- Removed unnecessary and redundancy in our Memtable iterators.
- The memtable iterator is now lazily evaluated
- All our store/lsm iterators now follow the nice principles we first learned from sstable iterators.
- A thinner lsm/store.rs core Public API

All of this makes the code so much easier to read and therefore easiier to document.
