+++ src/store/lsm/flush.rs
@@ -4,7 +4,7 @@ use crate::error::Result;
 /// Manually flush oldest frozen memtable to SSTable (for testing)
 pub async fn flush_memtable(store: &super::LsmStore) -> Result<bool> {
     // Check if flush needed and try to mark as pending
-    if !store.needs_flush() || !store.try_mark_flush_pending() {
+    if !store.needs_flush() || !store.state.try_mark_flush_pending() {
         return Ok(false);
     }
 
@@ -15,8 +15,7 @@ pub async fn flush_memtable(store: &super::LsmStore) -> Result<bool> {
         match frozen.pop_front() {
             Some(m) => m,
             None => {
-                // Mark flush completed and return
-                store.mark_flush_completed();
+                store.state.mark_flush_completed();
                 return Ok(false);
             }
         }
@@ -103,7 +102,7 @@ pub async fn flush_memtable(store: &super::LsmStore) -> Result<bool> {
     );
 
     // Mark flush as completed
-    store.mark_flush_completed();
+    store.state.mark_flush_completed();
 
     // Validate state consistency after flush
     if let Err(e) = store.state.validate_consistency() {
@@ -123,4 +122,181 @@ pub async fn flush_memtable(store: &super::LsmStore) -> Result<bool> {
     }
 
     Ok(true)
+}
+
+#[cfg(test)]
+mod tests {
+    use crate::{
+        config::{CompactionConfig, LsmConfig},
+        error::Result,
+        store::Store,
+        tmpfs::TempDir,
+    };
+
+    // Helper function to create a test store with custom config
+    fn create_test_store_with_config(
+        temp_dir: &TempDir,
+        config: CompactionConfig,
+    ) -> super::super::LsmStore {
+        let lsm_config = LsmConfig::new(temp_dir.path()).compaction(config);
+        super::super::LsmStore::open_with_config(lsm_config).expect("Failed to create store")
+    }
+
+    // Helper function to create a test store with default aggressive compaction settings
+    fn create_test_store(temp_dir: &TempDir) -> super::super::LsmStore {
+        let compaction_config = CompactionConfig::default()
+            .level0_compaction_threshold(2) // Lower threshold for easier testing
+            .size_ratio_threshold(2) // Lower ratio for easier testing
+            .max_tables_per_level(3); // Lower max for easier testing
+
+        create_test_store_with_config(temp_dir, compaction_config)
+    }
+
+    #[tokio::test]
+    async fn test_flush_memtable_basic() -> Result<()> {
+        let temp_dir = TempDir::new().expect("Failed to create temp dir");
+        let store = create_test_store(&temp_dir);
+
+        // Add some data to active memtable
+        for i in 0..10 {
+            let key = format!("key_{:03}", i);
+            let value = format!("value_{}", i);
+            store.set(key.as_bytes(), value.as_bytes().to_vec())?;
+        }
+
+        // Freeze the memtable to create a frozen one
+        store.freeze_active_memtable()?;
+
+        // Verify we have a frozen memtable
+        let frozen_count = {
+            let frozen = store.state.frozen_memtables.read().unwrap();
+            frozen.len()
+        };
+        assert_eq!(frozen_count, 1, "Should have one frozen memtable");
+
+        // Flush the memtable
+        let flushed = store.flush_memtable().await?;
+        assert!(flushed, "Should have flushed a memtable");
+
+        // Verify frozen memtable is gone
+        let frozen_count_after = {
+            let frozen = store.state.frozen_memtables.read().unwrap();
+            frozen.len()
+        };
+        assert_eq!(
+            frozen_count_after, 0,
+            "Should have no frozen memtables after flush"
+        );
+
+        // Verify data is still accessible
+        for i in 0..10 {
+            let key = format!("key_{:03}", i);
+            let value = store.get(key.as_bytes())?;
+            assert!(
+                value.is_some(),
+                "Key {} should be accessible after flush",
+                key
+            );
+        }
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_flush_memtable_empty() -> Result<()> {
+        let temp_dir = TempDir::new().expect("Failed to create temp dir");
+        let store = create_test_store(&temp_dir);
+
+        // Try to flush when no frozen memtables exist
+        let flushed = store.flush_memtable().await?;
+        assert!(!flushed, "Should not flush when no frozen memtables");
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_flush_memtable_multiple() -> Result<()> {
+        let temp_dir = TempDir::new().expect("Failed to create temp dir");
+        let store = create_test_store(&temp_dir);
+
+        // Create multiple frozen memtables
+        for batch in 0..3 {
+            for i in 0..5 {
+                let key = format!("batch_{}_key_{:03}", batch, i);
+                let value = format!("value_{}", i);
+                store.set(key.as_bytes(), value.as_bytes().to_vec())?;
+            }
+            store.freeze_active_memtable()?;
+        }
+
+        // Verify we have 3 frozen memtables
+        let frozen_count = {
+            let frozen = store.state.frozen_memtables.read().unwrap();
+            frozen.len()
+        };
+        assert_eq!(frozen_count, 3, "Should have three frozen memtables");
+
+        // Flush all memtables
+        let mut flush_count = 0;
+        while store.flush_memtable().await? {
+            flush_count += 1;
+        }
+        assert_eq!(flush_count, 3, "Should have flushed 3 memtables");
+
+        // Verify all data is still accessible
+        for batch in 0..3 {
+            for i in 0..5 {
+                let key = format!("batch_{}_key_{:03}", batch, i);
+                let value = store.get(key.as_bytes())?;
+                assert!(value.is_some(), "Key {} should be accessible", key);
+            }
+        }
+
+        Ok(())
+    }
+
+    #[tokio::test]
+    async fn test_manual_flush_cycle() -> Result<()> {
+        let temp_dir = TempDir::new().expect("Failed to create temp dir");
+        let store = create_test_store(&temp_dir);
+
+        // Initially should not need flush
+        assert!(!store.needs_flush(), "Fresh store should not need flush");
+
+        // Write data until we have a frozen memtable
+        for i in 0..10 {
+            let key = format!("key_{:03}", i);
+            let value = format!("value_{}", i);
+            store.set(key.as_bytes(), value.as_bytes().to_vec())?;
+        }
+
+        // Manually freeze to create a frozen memtable
+        store.freeze_active_memtable()?;
+
+        // Now should need flush
+        assert!(
+            store.needs_flush(),
+            "Should need flush after freezing memtable"
+        );
+
+        // Manual flush using public API
+        let flushed = store.flush_memtable().await?;
+        assert!(flushed, "Should have flushed a memtable");
+
+        // Should no longer need flush
+        assert!(!store.needs_flush(), "Should not need flush after flushing");
+
+        // Verify data is still accessible
+        for i in 0..10 {
+            let key = format!("key_{:03}", i);
+            let value = store.get(key.as_bytes())?;
+            assert!(
+                value.is_some(),
+                "Key {} should be accessible after flush",
+                key
+            );
+        }
+
+        Ok(())
+    }
 }
