+++ src/store/lsm/iterator.rs
@@ -1,133 +1,103 @@
 use crate::error::Result;
-use std::{cmp::Ordering, collections::BinaryHeap, ops::RangeBounds, sync::Arc};
+use std::{cmp::Ordering, collections::BinaryHeap};
 
 /// Type alias for complex iterator used in merge operations
-pub type KvIterator<'a> = Box<dyn Iterator<Item = Result<(Vec<u8>, Vec<u8>)>> + Send + Sync + 'a>;
+pub type LSMIterator<'a> = Box<dyn Iterator<Item = Result<(Vec<u8>, Vec<u8>)>> + Send + Sync + 'a>;
 
-// Iterator implementations
-pub struct OwningMemtableIter {
-    _memtable: Arc<super::memtable::Memtable>,
-    iter: super::memtable::ScanIter<'static>,
+/// Simplified heap entry that doesn't store the iterator inside.
+/// The iterator is kept in an external Vec and referenced by index.
+#[derive(Debug)]
+pub struct HeapEntry {
+    key: Vec<u8>,
+    value: Vec<u8>,
+    source_index: usize,
 }
 
-impl OwningMemtableIter {
-    pub fn new(memtable: Arc<super::memtable::Memtable>, range: impl RangeBounds<Vec<u8>>) -> Self {
-        let iter = unsafe {
-            std::mem::transmute::<super::memtable::ScanIter<'_>, super::memtable::ScanIter<'static>>(
-                memtable.scan(range).unwrap(),
-            )
-        };
-
-        Self {
-            _memtable: memtable,
-            iter,
-        }
-    }
-}
-
-impl Iterator for OwningMemtableIter {
-    type Item = Result<(Vec<u8>, Vec<u8>)>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.iter.next()
-    }
-}
-
-pub struct HeapEntry<'a> {
-    pub key: Vec<u8>,
-    pub value: Vec<u8>,
-    pub source: usize,
-    pub iterator: KvIterator<'a>,
-}
-
-impl std::fmt::Debug for HeapEntry<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("HeapEntry")
-            .field("key", &self.key)
-            .field("value", &self.value)
-            .field("source", &self.source)
-            .finish()
-    }
-}
-
-impl PartialEq for HeapEntry<'_> {
+impl PartialEq for HeapEntry {
     fn eq(&self, other: &Self) -> bool {
         self.key == other.key
     }
 }
 
-impl Eq for HeapEntry<'_> {}
+impl Eq for HeapEntry {}
 
-impl PartialOrd for HeapEntry<'_> {
+impl PartialOrd for HeapEntry {
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         Some(self.cmp(other))
     }
 }
 
-impl Ord for HeapEntry<'_> {
+impl Ord for HeapEntry {
     fn cmp(&self, other: &Self) -> Ordering {
         match self.key.cmp(&other.key) {
-            Ordering::Equal => self.source.cmp(&other.source),
-            other => other.reverse(),
+            Ordering::Equal => self.source_index.cmp(&other.source_index),
+            other => other.reverse(), // Reverse for min-heap behavior
         }
     }
 }
 
-#[derive(Debug)]
-pub struct MergeIterator<'a> {
-    heap: BinaryHeap<HeapEntry<'a>>,
-    latest_key: Option<Vec<u8>>,
+/// Cleaned up LSM scan iterator that merges multiple sorted iterators.
+/// This maintains the same merge sort algorithm but with simpler implementation.
+pub struct LSMScanIterator<'a> {
+    iterators: Vec<LSMIterator<'a>>,
+    heap: BinaryHeap<HeapEntry>,
+    last_yielded_key: Option<Vec<u8>>,
 }
 
-impl<'a> MergeIterator<'a> {
-    pub fn new(iterators: Vec<KvIterator<'a>>) -> Self {
+impl<'a> LSMScanIterator<'a> {
+    pub fn new(mut iterators: Vec<LSMIterator<'a>>) -> Self {
         let mut heap = BinaryHeap::new();
 
-        for (source, mut iterator) in iterators.into_iter().enumerate() {
+        // Prime the heap with the first entry from each iterator
+        for (source_index, iterator) in iterators.iter_mut().enumerate() {
             if let Some(Ok((key, value))) = iterator.next() {
                 heap.push(HeapEntry {
                     key,
                     value,
-                    source,
-                    iterator,
+                    source_index,
                 });
             }
         }
 
         Self {
+            iterators,
             heap,
-            latest_key: None,
+            last_yielded_key: None,
         }
     }
 }
 
-impl Iterator for MergeIterator<'_> {
+impl Iterator for LSMScanIterator<'_> {
     type Item = Result<(Vec<u8>, Vec<u8>)>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        while let Some(mut entry) = self.heap.pop() {
-            if self.latest_key.as_ref() == Some(&entry.key) {
-                if let Some(Ok((key, value))) = entry.iterator.next() {
+        while let Some(entry) = self.heap.pop() {
+            // Skip duplicates - if we see the same key again, newer source wins
+            if self.last_yielded_key.as_ref() == Some(&entry.key) {
+                // Pull the next entry from this iterator to keep it going
+                if let Some(Ok((key, value))) = self.iterators[entry.source_index].next() {
                     self.heap.push(HeapEntry {
                         key,
                         value,
-                        source: entry.source,
-                        iterator: entry.iterator,
+                        source_index: entry.source_index,
                     });
                 }
-                continue;
+                continue; // Skip this duplicate key
             }
 
-            self.latest_key = Some(entry.key.clone());
+            // Update our last yielded key for deduplication
+            self.last_yielded_key = Some(entry.key.clone());
 
-            if let Some(Ok((key, value))) = entry.iterator.next() {
+            // Pull the next entry from this iterator
+            if let Some(Ok((key, value))) = self.iterators[entry.source_index].next() {
                 self.heap.push(HeapEntry {
                     key,
                     value,
-                    source: entry.source,
-                    iterator: entry.iterator,
+                    source_index: entry.source_index,
                 });
             }
+
+            // Return this entry
             return Some(Ok((entry.key, entry.value)));
         }
