+++ src/store/lsm/sstable/mod.rs
@@ -39,5 +39,4 @@
 mod index;
 
 pub mod block;
-pub mod filter;
 pub mod table;
